+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/buildin/hardmediumsoftlong/HardMediumSoftLongScoreHolderImpl.java
@@ -16,6 +16,7 @@
 
 package org.optaplanner.core.impl.score.buildin.hardmediumsoftlong;
 
+import java.math.BigDecimal;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.function.BiConsumer;
@@ -149,6 +150,14 @@ public final class HardMediumSoftLongScoreHolderImpl extends AbstractScoreHolder
                     + ConstraintConfiguration.class.getSimpleName() + " annotated class.");
         }
         matchExecutor.accept(kcontext, weightMultiplier);
+    }
+
+    @Override
+    public void impactScore(RuleContext kcontext, BigDecimal weightMultiplier) {
+        throw new UnsupportedOperationException("In the rule (" + kcontext.getRule().getName()
+                + "), the scoreHolder class (" + getClass()
+                + ") does not support a BigDecimal weightMultiplier (" + weightMultiplier + ").\n"
+                + "If you're using constraint streams, maybe switch from penalizeBigDecimal() to penalizeLong()?");
     }
 
     private void impactScore(RuleContext kcontext, long hardWeightMultiplier, long mediumWeightMultiplier,
